GPU (floating decimal)

+++ java/jupiter/integration/gpu/resources/OpenCL/program.cl
@@ -1,4 +1,4 @@
-__kernel void plus(__global const double* A, __global const double* B, __global double* C) {
+__kernel void plus(__global const float* A, __global const float* B, __global float* C) {
 	// Find the index in the global arrays
 	const int index = get_global_id(0);
 
@@ -6,7 +6,7 @@ __kernel void plus(__global const double* A, __global const double* B, __global 
 	C[index] = A[index] + B[index];
 }
 
-__kernel void minus(__global const double* A, __global const double* B, __global double* C) {
+__kernel void minus(__global const float* A, __global const float* B, __global float* C) {
 	// Find the index in the global arrays
 	const int index = get_global_id(0);
 
@@ -14,7 +14,7 @@ __kernel void minus(__global const double* A, __global const double* B, __global
 	C[index] = A[index] - B[index];
 }
 
-__kernel void times(__global const double* A, __global const double* B, __global double* C,
+__kernel void times(__global const float* A, __global const float* B, __global float* C,
 		const int aColumnDimension, const int bColumnDimension) {
 	// Find the index in the global arrays
 	const int index = get_global_id(0);
@@ -24,14 +24,14 @@ __kernel void times(__global const double* A, __global const double* B, __global
 	const int columnOffset = index % bColumnDimension;
 
 	// Compute the dot product
-	double sum = 0.0;
+	float sum = 0.0;
 	for (int i = 0; i < aColumnDimension; ++i) {
 		sum += A[rowOffset * aColumnDimension + i] * B[i * bColumnDimension + columnOffset];
 	}
 	C[index] = sum;
 }
 
-__kernel void arrayTimes(__global const double* A, __global const double* B, __global double* C) {
+__kernel void arrayTimes(__global const float* A, __global const float* B, __global float* C) {
 	// Find the index in the global arrays
 	const int index = get_global_id(0);
 
@@ -39,7 +39,7 @@ __kernel void arrayTimes(__global const double* A, __global const double* B, __g
 	C[index] = A[index] * B[index];
 }
 
-__kernel void arrayDivision(__global const double* A, __global const double* B, __global double* C) {
+__kernel void arrayDivision(__global const float* A, __global const float* B, __global float* C) {
 	// Find the index in the global arrays
 	const int index = get_global_id(0);
 
@@ -47,7 +47,7 @@ __kernel void arrayDivision(__global const double* A, __global const double* B, 
 	C[index] = A[index] / B[index];
 }
 
-__kernel void arrayLeftDivision(__global const double* A, __global const double* B, __global double* C) {
+__kernel void arrayLeftDivision(__global const float* A, __global const float* B, __global float* C) {
 	// Find the index in the global arrays
 	const int index = get_global_id(0);
